extract common stuff
